Struct openai::chat::ChatCompletionBuilder

pub struct ChatCompletionBuilder { /* private fields */ }

Builder for ChatCompletionRequest.

Implementations

impl ChatCompletionBuilder

pub fn model<VALUE: Into<ModelID>>(self, value: VALUE) -> Self

ID of the model to use. Currently, only gpt-3.5-turbo and gpt-3.5-turbo-0301 are supported.

pub fn messages<VALUE: Into<Vec<ChatCompletionMessage>>>( self, value: VALUE ) -> Self

The messages to generate chat completions for, in the chat format.

pub fn temperature<VALUE: Into<f32>>(self, value: VALUE) -> Self

What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.

We generally recommend altering this or top_p but not both.

pub fn top_p<VALUE: Into<f32>>(self, value: VALUE) -> Self

An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.

We generally recommend altering this or temperature but not both.

pub fn n<VALUE: Into<u8>>(self, value: VALUE) -> Self

How many chat completion choices to generate for each input message.

pub fn stop<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

Up to 4 sequences where the API will stop generating further tokens.

pub fn max_tokens<VALUE: Into<u64>>(self, value: VALUE) -> Self

The maximum number of tokens allowed for the generated answer. By default, the number of tokens the model can return will be (4096 - prompt tokens).

pub fn presence_penalty<VALUE: Into<f32>>(self, value: VALUE) -> Self

Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model’s likelihood to talk about new topics.

See more information about frequency and presence penalties.

pub fn frequency_penalty<VALUE: Into<f32>>(self, value: VALUE) -> Self

Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model’s likelihood to repeat the same line verbatim.

See more information about frequency and presence penalties.

pub fn logit_bias<VALUE: Into<HashMap<String, f32>>>(self, value: VALUE) -> Self

Modify the likelihood of specified tokens appearing in the completion.

Accepts a json object that maps tokens (specified by their token ID in the tokenizer) to an associated bias value from -100 to 100. Mathematically, the bias is added to the logits generated by the model prior to sampling. The exact effect will vary per model, but values between -1 and 1 should decrease or increase likelihood of selection; values like -100 or 100 should result in a ban or exclusive selection of the relevant token.

pub fn user<VALUE: Into<String>>(self, value: VALUE) -> Self

A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse. Learn more.

pub fn build(self) -> Result<ChatCompletionRequest, ChatCompletionBuilderError>

Builds a new ChatCompletionRequest.

Errors

If a required field has not been initialized.

impl ChatCompletionBuilder

pub async fn create(self) -> Result<Result<ChatCompletion, OpenAiError>, Error>