Struct openai::completions::CompletionBuilder

pub struct CompletionBuilder { /* private fields */ }

Builder for CompletionRequest.

Implementations

impl CompletionBuilder

pub fn model<VALUE: Into<ModelID>>(self, value: VALUE) -> Self

ID of the model to use. You can use the List models API to see all of your available models, or see our Model overview for descriptions of them.

pub fn prompt<VALUE: Into<String>>(self, value: VALUE) -> Self

The prompt(s) to generate completions for, encoded as a string, array of strings, array of tokens, or array of token arrays.

Note that <|endoftext|> is the document separator that the model sees during training, so if a prompt is not specified the model will generate as if from the beginning of a new document.

pub fn suffix<VALUE: Into<String>>(self, value: VALUE) -> Self

The suffix that comes after a completion of inserted text.

pub fn max_tokens(self, value: u16) -> Self

The maximum number of tokens to generate in the completion.

The token count of your prompt plus max_tokens cannot exceed the model’s context length. Most models have a context length of 2048 tokens (except for the newest models, which support 4096).

pub fn temperature<VALUE: Into<f32>>(self, value: VALUE) -> Self

What sampling temperature to use. Higher values means the model will take more risks. Try 0.9 for more creative applications, and 0 (argmax sampling) for ones with a well-defined answer.

We generally recommend altering this or top_p but not both.

pub fn top_p<VALUE: Into<f32>>(self, value: VALUE) -> Self

An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.

We generally recommend altering this or temperature but not both.

pub fn n<VALUE: Into<u16>>(self, value: VALUE) -> Self

How many completions to generate for each prompt.

Note: Because this parameter generates many completions, it can quickly consume your token quota. Use carefully and ensure that you have reasonable settings for max_tokens and stop.

pub fn logprobs<VALUE: Into<u8>>(self, value: VALUE) -> Self

Include the log probabilities on the logprobs most likely tokens, as well the chosen tokens. For example, if logprobs is 5, the API will return a list of the 5 most likely tokens. The API will always return the logprob of the sampled token, so there may be up to logprobs+1 elements in the response.

The maximum value for logprobs is 5. If you need more than this, please contact us through our Help center and describe your use case.

pub fn echo<VALUE: Into<bool>>(self, value: VALUE) -> Self

Echo back the prompt in addition to the completion

pub fn stop<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

Up to 4 sequences where the API will stop generating further tokens. The returned text will not contain the stop sequence.

pub fn presence_penalty<VALUE: Into<i8>>(self, value: VALUE) -> Self

Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model’s likelihood to talk about new topics.

See more information about frequency and presence penalties.

pub fn frequency_penalty<VALUE: Into<i8>>(self, value: VALUE) -> Self

Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model’s likelihood to repeat the same line verbatim.

See more information about frequency and presence penalties.

pub fn best_of<VALUE: Into<u16>>(self, value: VALUE) -> Self

Generates best_of completions server-side and returns the “best” (the one with the highest log probability per token). Results cannot be streamed.

When used with n, best_of controls the number of candidate completions and n specifies how many to return – best_of must be greater than n.

Note: Because this parameter generates many completions, it can quickly consume your token quota. Use carefully and ensure that you have reasonable settings for max_tokens and stop.

pub fn logit_bias<VALUE: Into<HashMap<String, i16>>>(self, value: VALUE) -> Self

Modify the likelihood of specified tokens appearing in the completion.

Accepts a json object that maps tokens (specified by their token ID in the GPT tokenizer) to an associated bias value from -100 to 100. You can use this tokenizer tool (which works for both GPT-2 and GPT-3) to convert text to token IDs. Mathematically, the bias is added to the logits generated by the model prior to sampling. The exact effect will vary per model, but values between -1 and 1 should decrease or increase likelihood of selection; values like -100 or 100 should result in a ban or exclusive selection of the relevant token.

As an example, you can pass {"50256": -100} to prevent the <|endoftext|> token from being generated.

pub fn user<VALUE: Into<String>>(self, value: VALUE) -> Self

A unique identifier representing your end-user, which can help OpenAI to monitor and detect abuse. Learn more.

pub fn build(self) -> Result<CompletionRequest, CompletionBuilderError>

Builds a new CompletionRequest.

Errors

If a required field has not been initialized.

impl CompletionBuilder

pub async fn create(self) -> Result<Result<Completion, OpenAiError>, Error>